api_service/pyproject.toml
[project]
name = "ballskicker-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioboto3>=13.4.0",
    "aiocache>=0.12.3",
    "boto3>=1.36.0",
    "boto3-stubs[dynamodb,s3]>=1.36.2",
    "cachetools>=5.5.2",
    "fastapi>=0.115.6",
    "httpx>=0.28.1",
    "jwt>=1.3.1",
    "pydantic-settings>=2.7.1",
    "pydantic[email]>=2.10.5",
    "python-jose>=3.3.0",
    "types-aioboto3>=13.4.0",
    "uvicorn>=0.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ballskicker_api"]

[dependency-groups]
dev = [
    "mypy>=1.14.1",
    "pre-commit>=4.1.0",
    "ruff>=0.9.2",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP"]
ignore = ["E731"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
